UI for the Model
# Diabetes Prediction Project
The Diabetes Prediction Project is a web application that predicts the likelihood of a person having diabetes based on various factors. It utilizes a trained machine learning model to provide accurate predictions. The project includes two files: index.html, which contains the web page code, and app.py, which contains the Flask server code and the implementation of the trained model.

Features
------------
Predicts the likelihood of a person having diabetes based on input factors.
Provides a user-friendly web interface for entering the required information.
Displays the prediction result on the web page.
Utilizes a trained machine learning model to make accurate predictions.

Technologies Used
----------------------
Python,
Flask,
HTML,
CSS,
Machine Learning

Repository Structure
------------------------
The repository for the Diabetes Prediction Project contains the following files:

index.html: This file contains the HTML code for the web page. It includes the form for entering the input data and displays the prediction result.

app.py: This file contains the Flask server code. It handles the HTTP requests, processes the input data, and uses the trained machine learning model to make predictions.

Trained Model File: The trained machine learning model file is not included in the repository. It should be trained separately and loaded within the app.py file.

Deployment
--------------
The Diabetes Prediction Project is currently deployed on PythonAnywhere. You can access the web application at rudra141.pythonanywhere.com.

Contributing
--------------
Contributions to the Diabetes Prediction Project are welcome. If you find any issues or have suggestions for improvements, feel free to submit a pull request or open an issue in the project repository.

License
-------------
The Diabetes Prediction Project is released under the MIT License. See the LICENSE file for more details.

Acknowledgments
-----------------
The Diabetes Prediction Project acknowledges the use of machine learning algorithms and techniques to create an accurate prediction model for diabetes. Special thanks to the contributors of the open-source libraries used in this project.